% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticky_tbl_df.R
\name{as_sticky_tibble}
\alias{as_sticky_tibble}
\title{Coerce objects to sticky tibble}
\usage{
as_sticky_tibble(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused, for extensibility.}
}
\value{
A sticky tibble ('sticky_tbl_df') object.
}
\description{
'as_sticky_tibble' turns an object to a sticky tibble ('sticky_tbl_df').
}
