% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticky_tbl_df.R
\name{is_sticky_tibble}
\alias{is_sticky_tibble}
\title{Test if the object is a sticky tibble}
\usage{
is_sticky_tibble(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if an object inherits from \code{sticky_tbl_df},
\code{sticky_grouped_df} or \code{sticky_rowwise_df} class.
}
\description{
Test if the object is a sticky tibble
}
