% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticky_tbl_df.R
\name{new_sticky_tibble}
\alias{new_sticky_tibble}
\title{Constructs a tibble with persistent columns and attributes}
\usage{
new_sticky_tibble(
  x = list(),
  cols = NULL,
  col_show = tidyselect::everything(),
  col_summary = list(),
  attrs = character(),
  ...,
  class = character(),
  class_grouped_df = class,
  class_rowwise_df = class
)
}
\arguments{
\item{x}{A tibble-like object.}

\item{cols}{Columns to be kept persistently.}

\item{col_show}{Which persistent columns to show? By default, all are
displayed.}

\item{col_summary}{A named list of functions to summarise persistent columns.}

\item{attrs}{Names of attributes to be kept persistently.}

\item{...}{Additional attributes.}

\item{class}{Subclasses.}

\item{class_grouped_df}{Subclasses of grouped_df. By default, it is the same
as \code{class}.}

\item{class_rowwise_df}{Subclasses of rowwise_df. By default, it is the same
as \code{class}.}
}
\value{
A sticky tibble (\code{sticky_tbl_df}) object.
}
\description{
\code{new_sticky_tibble} constructs a tibble with persistent columns and
attributes. These persistent columns can be hidden.
}
