% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_main.R
\name{select.tbl_main}
\alias{select.tbl_main}
\title{#' @importFrom dplyr rename
#' @export
rename.tbl_main <- function(.data, ...) {
loc <- tidyselect::eval_rename(expr(c(...)), .data)
names <- names(.data)
names\link{loc} <- names(loc)
.data <- set_names(.data, names)}
\usage{
\method{select}{tbl_main}(.data, ...)
}
\description{

}
