% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R, R/stickyr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{intersect}
\alias{setdiff}
\alias{union}
\alias{tbl_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{generics}{\code{\link[generics:setops]{intersect}}, \code{\link[generics:setops]{setdiff}}, \code{\link[generics:setops]{union}}}

  \item{pillar}{\code{\link[pillar]{tbl_sum}}}
}}

